requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0

certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
dateutils==0.6.12
gunicorn
django-silk
defusedxml==0.7.1
Django==4.0.5
django-extensions==3.1.5
django-filter==22.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-user-agents==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
autopep8
flake8
drf-yasg==1.20.0
frozenlist==1.3.0
geoip2==4.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
maxminddb==2.2.0
multidict==6.0.2
mysqlclient==2.1.1
oauthlib==3.2.0
packaging==21.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
whitenoise
django-cors-headers
django-heroku
redis
celery
django-celery-beat
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
ua-parser==0.15.0
uritemplate==4.1.1
urllib3==1.26.9
user-agents==2.2.0
yarl==1.7.2
